src/requirements.txt
cryptography
python-dotenv 
fastapi